requirements.txt
flask==3.1.*
gunicorn==23.*
pyjwt[crypto]==2.9.*
